Ex 6.7: anecdotes, step 5. Separated creation of new anecdotes into component <AnecdoteForm />. Moved logic for creating new anecdote into <AnecdoteForm /> component.
