+++ submissions/anecdotes-redux/src/App.js
@@ -1,19 +1,13 @@
 import React from 'react'
+import AnecdoteForm from './components/AnecdoteForm'
 import { useSelector, useDispatch } from 'react-redux'
-import { voteForAnecdote, createAnecdote, getId } from './reducers/anecdoteReducer'
+import { voteForAnecdote, createAnecdote } from './reducers/anecdoteReducer'
 
 const App = () => {
   const anecdotes = useSelector(state => state)
- 
   const dispatch = useDispatch()
 
   const vote = (id) => dispatch(voteForAnecdote(id))
-
-  const addAnecdote = event => {
-    event.preventDefault()
-    dispatch(createAnecdote(event.target.anecdote.value))
-    event.target.anecdote.value = ''
-  }
 
   return (
     <div>
@@ -29,11 +23,8 @@ const App = () => {
           </div>
         </div>
       )}
-      <h2>create new</h2>
-      <form onSubmit={addAnecdote}>
-        <div><input name="anecdote" /></div>
-        <button type="submit">create</button>
-      </form>
+      <AnecdoteForm dispatch={dispatch} />
+
     </div>
   )
 }
